refactor: move export_address to BTC-specific package
- Move export_address.go from shared/ to btc/ package
- Remove multi-coin logic (ETH, XRP) as this implementation only works with BTCAccountKeyRepositorier
- Add validation to explicitly reject non-BTC/BCH coin types
- Update DI container to use BTC-specific implementation
- Add documentation noting that ETH/XRP require separate implementations

This addresses the architectural issue identified in review where the use case
claimed to support multiple coin types but only worked with BTC repository,
causing silent failures for ETH/XRP addresses.

Fixes: Review comment on PR #272
